Better support for multi-manager downloads.
This is a rewrite of SDWebImageDownloader's operation bookkeeping. 

SDWebImageDownloader previously only created one download operation and returned it the first time the URL was requested. If additional requests for the same URL were made it added the callbacks but did not return anything the caller. 

This is problematic when we have two or more managers querying for the same URL (e.g, the default manager and a manager inside the prefetcher). The first manager gets a reference to the operation while the other managers get a nil response. This causes:
 - cancelation calls on the first manager to also cancel downloads for all other managers 
 - cancelation calls on subsequent managers to have no effect at all (since they have no reference to the download operation)

This rewrite fixes those issues. Each time an URL is requested a SDWebImageDowloaderObserverOperation is created and returned. SDWebImageDowloaderObserverOperation contains a reference to the download operation and references to various download callbacks. Download requests for the same URL get a different SDWebImageDowloaderObserverOperation, but share the same download operation inside.

When a SDWebImageDowloaderObserverOperation gets cancelled, the observation is removed and the callbacks inside of it will no longer be called. If happened to also be the last observer for a specific URL, the download gets cancelled as well. This essentially reference counts interested observers for an active download. If the reference count gets to zero, the download is cancelled. 

This is especially useful if we use one manager (the default manager) to load images inside table view cells and a preloader to load additional cells when the user is not scrolling. As soon as the user resumes scrolling, we can cancel the prefetcher downloads, whiteout having to worry that this might also cancel downloads inside visible cells.